requirements.txt
flask-cors
Flask-SQLAlchemy==2.1
coverage
pylint
Flask
flask-socketio
psycopg2-binary
flask-marshmallow==0.14.0
marshmallow==3.11.1
marshmallow-sqlalchemy==0.24.3